Experiementing with JSON xpath support
