+++ rest-assured/pom.xml
@@ -40,6 +40,21 @@
             <artifactId>hamcrest-core</artifactId>
             <version>1.1</version>
         </dependency>
+        <dependency>
+            <groupId>commons-jxpath</groupId>
+            <artifactId>commons-jxpath</artifactId>
+            <version>1.3</version>
+        </dependency>
+        <dependency>
+            <groupId>org.codehaus.jackson</groupId>
+            <artifactId>jackson-mapper-asl</artifactId>
+            <version>1.6.2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.powermock</groupId>
+            <artifactId>powermock-reflect</artifactId>
+            <version>1.4.6</version>
+        </dependency>
 
         <!-- Test dependencies -->
         <dependency>
